Two programs, one of it have additional feature.
The program names ForensicURL.py have an additional feature of generating an PDF file of the result which is generated by the program. The PDF is useful for future uses. The generation of PDF file is automatically initiated when the analysis executed.   NOTE: The generating speed rate of PDF is depends on your network speed and system performance.